Create Stacks By list
A stack is used to store data 
such that the last item 
inserted is the first item 
removed. 
The stack is a linear data 
structure which implements 
the last-in-first-out (LIFO) 
protocol. 
New items are added, and 
existing items are removed 
from the same end, called the 
top of the stack (TOS).